+++ src/hooks/useDatasets.js
@@ -1,10 +1,10 @@
-import { useState, useEffect, useCallback, useMemo } from "react";
+import { useState, useEffect, useCallback } from "react";
 import { parse } from "papaparse";
 
-import { BASE_URL, DATASETS } from "./constants";
-import { formatDataSets } from "./utils";
+import { BASE_URL, DATASETS } from "../constants";
+import { formatDataSets } from "../utils";
 
-export const useDatasets = () => {
+const useDatasets = () => {
   const [data, setData] = useState(null);
   const [error, setError] = useState(null);
 
@@ -53,22 +53,4 @@ export const useDatasets = () => {
   return [data, error];
 };
 
-export const useDatasetsLimits = (data) => {
-  const limits = useMemo(() => {
-    let max = -Infinity;
-    let min = Infinity;
-    Object.values(data).forEach((year) => {
-        return Object.values(year).forEach((country) =>
-          Object.values(country).forEach((val) => {
-            if (val) {
-              max = Math.max(val, max);
-              min = Math.min(val, min);
-            }
-          })
-        );
-      });
-    return [max, min];
-  }, [data]);
-
-  return limits;
-};
+export default useDatasets;
